refactor: reorgranise session access classes
